+++ ent/hymnswork/where.go
@@ -11,47 +11,47 @@ import (
 )
 
 // ID filters vertices based on their ID field.
-func ID(id int) predicate.HymnsWork {
+func ID(id int64) predicate.HymnsWork {
 	return predicate.HymnsWork(sql.FieldEQ(FieldID, id))
 }
 
 // IDEQ applies the EQ predicate on the ID field.
-func IDEQ(id int) predicate.HymnsWork {
+func IDEQ(id int64) predicate.HymnsWork {
 	return predicate.HymnsWork(sql.FieldEQ(FieldID, id))
 }
 
 // IDNEQ applies the NEQ predicate on the ID field.
-func IDNEQ(id int) predicate.HymnsWork {
+func IDNEQ(id int64) predicate.HymnsWork {
 	return predicate.HymnsWork(sql.FieldNEQ(FieldID, id))
 }
 
 // IDIn applies the In predicate on the ID field.
-func IDIn(ids ...int) predicate.HymnsWork {
+func IDIn(ids ...int64) predicate.HymnsWork {
 	return predicate.HymnsWork(sql.FieldIn(FieldID, ids...))
 }
 
 // IDNotIn applies the NotIn predicate on the ID field.
-func IDNotIn(ids ...int) predicate.HymnsWork {
+func IDNotIn(ids ...int64) predicate.HymnsWork {
 	return predicate.HymnsWork(sql.FieldNotIn(FieldID, ids...))
 }
 
 // IDGT applies the GT predicate on the ID field.
-func IDGT(id int) predicate.HymnsWork {
+func IDGT(id int64) predicate.HymnsWork {
 	return predicate.HymnsWork(sql.FieldGT(FieldID, id))
 }
 
 // IDGTE applies the GTE predicate on the ID field.
-func IDGTE(id int) predicate.HymnsWork {
+func IDGTE(id int64) predicate.HymnsWork {
 	return predicate.HymnsWork(sql.FieldGTE(FieldID, id))
 }
 
 // IDLT applies the LT predicate on the ID field.
-func IDLT(id int) predicate.HymnsWork {
+func IDLT(id int64) predicate.HymnsWork {
 	return predicate.HymnsWork(sql.FieldLT(FieldID, id))
 }
 
 // IDLTE applies the LTE predicate on the ID field.
-func IDLTE(id int) predicate.HymnsWork {
+func IDLTE(id int64) predicate.HymnsWork {
 	return predicate.HymnsWork(sql.FieldLTE(FieldID, id))
 }
